drivers: Add library for reconfiguring Xilinx high-speed transceiver
The Xilinx high-speed transceivers feature a so called
dynamic-reconfiguration-port (DRP) that can be used to reconfigure the
operating parameters of the transceiver at runtime.

Add a standalone library module that can be used to calculate and
program the DRP settings that are relevant for JESD204 operation. It
can be used by drivers that need to re-program the high-speed
transceivers to handle the low level details.

Signed-off-by: Dragos Bogdan <dragos.bogdan@analog.com>