projects: ad9371: Add an example for using DAC DMA
Commit 4326a42b5 offers an example of how to send custom data using the
DMA and adrv9009.
This patch updates the headless file in the same manner in order to work
with the ad9371 project. The main difference is that the configuration
for ad9371 uses the FIFO, while adrv9009 bypasses it. At the moment, the
DMA cyclic mode is not supported in the hdl.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>