jesd204: xcvr: Avoid using DRP broadcast to prevent undesired behaviour
The DRP BROADCAST feature will write the same value to all channels,
however, there are some cases when not all channels are configured in the
same way (e.g. the clock source is different). In order not to overwrite
other settings, the update to some registers should be made
independently for each channel. Also, the BROADCAST read might fail
since the result is OR-ed between all channels.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>