drivers:platform:xilinx: fix compile switches
The current I2C implementation is for the AXI IIC core. As such it would
make more sense to use the inclusion guard of the driver as a compile
switch instead of the inclusion guard of another, potentially unrelated,
header. This can cause problems if the current header is present, but with
no Xilinx AXI IIC core in the design.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>