adrv9009: always use pl-fifo for high speed dma
This commit fixes a bug that had as symptom ugly sinewaves on rx
channels. After thorough investigation, it seems the tx DMA was set
to cyclic and was overwriting the content of pl-fifo. With these
changes, we make sure tx DMA only writes the content of pl-fifo once
and then the pl-fifo ensures the data is sent onwards to the DAC.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>