libraries:iio: Added iio channel type indexing.
The indexed field is used to check if a number (the channel field) will
be attached to the channel type. E.g. altvoltage0, altvoltage1, ... The
reserved field in the channel is no longer used.

This commit affects the way the channel is matched during read/write and
during the generation of xml.

Signed-off-by: Jimenez, Kister <kister.jimenez@analog.com>