projects/ad9371: use buffer for DAC_DMA_EXAMPLE
DAC_DMA_EXAMPLE was previously using an arbitrary memory to store
the ADC and DAC data. The two addresses were set pretty high up in
the memory but this still does not guarantee that the area does
not get overwritten by the program because the size of the program
is not known in advance and no checks are performed.

The ADC and DAC buffers are then allocated statically within the
program and the address of the ADC buffer is printed to the console
when the DAC_DMA_EXAMPLE is enabled:

DAC_DMA_EXAMPLE: address=0x7f170 samples=65536 channels=4 bits=16

This information can be then used to fill in the parameters
xilinx_capture.tcl expects:

xsct xilinx_capture.tcl ZYNQ_PSU 0x7f170 65536 4 16

Signed-off-by: Darius Berghe <darius.berghe@analog.com>