projects:eval-adxl355-pmdz: Create EVAL-ADXL355-PMDZ project
The EVAL-ADXL355-PMDZ is a simple evaluation boards that allows
quick evaluation of the performance of the ADXL355 low noise,
low power, 3-axis, MEMS accelerometer.
The EVAL-ADXL355-PMDZ is a digital output supporting ±2 g, ±4 g,
or ±8 g. The evaluation board is ideal for evaluating the ADXL355 in
an existing system  because the stiffness and small size of the
evaluation board minimizes the effect of the evaluation board on both
the system  and acceleration measurements.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>