axi_adxcvr.* Addition of defines for out_clk_sel
The addition of defines for selecting the transceiver reference clock
as the source of TXOUTCLK.

The update of sample projects to use the selection of the
transceiver reference clock defines from file axi_adxcvr.h.

Signed-off-by: George Mois <george.mois@analog.com>

src:app:fmcdaq2.c Update of out clock selection

The update of the fmcdaq2 sample project to use the selection of the
transceiver reference clock defines from file axi_adxcvr.h.

Signed-off-by: George Mois <george.mois@analog.com>