no_os_uart_stdio common API across platforms
This functionality was scattered around platforms under different
names. Fix that, use a common name.

Because this is an optional feature, no_os_uart.c imlplements an
empty version with weak linkage. This means that platforms that
implement this feature override the weak function whereas the
rest of them can simply not implement it.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>