projects: adin1110: Add ADIN1110 example project
Add an example project which showcases the ethernet frames receive and
transmit functionality. Every 5 seconds a frame (payload filled with 0x44)
is broadcasted, and incoming frames are read (the device is set in
promiscuous mode, so there are no mac filters).

This is implemented for the ADIN2111. The targets are MAX32650 and
STM32F7.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>