projects: eval-adxl355pmdz: IIO networking example
Add an example of using the ADXL355 with the IIOD network backend
(+adin1110). This uses the ADIN1110EBZ as a carrier, and as such, the
SPI interface parameters (SPI id, CS port and number) had to be changed.
Should only affect the IIO_LWIP_EXAMPLE version of the project. For all
the other examples, the configuration is still done for SDP-K1.

Explicitly set the HARDWARE variable to the current SDP-K1 ioc file when
building the previous examples.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>