add eeprom / factory defaults functionality
There is an EEPROM on the board and we use it to keep in it specific
parameters that should be persistent across reboots.

This commit introduces the `factory defaults` concept and implements
the eeprom handling.

Parameters are stored in a `struct nvmp`, they are saved at address
0 of the eeprom, and after them, a crc8 byte is saved as well.

Detecting a bad crc makes the firmware replace the values already in
eeprom with factory defaults.

Detecting a good crc makes the firmware use the values from the eeprom
when it performs initializations.

The PC application has a button that allows the user to save parameters
manually into the EEPROM. They are stored in bulk, the complete subset
is stored upon clicking that button. This functionality is possible
via a new attribute of `mwc` called `save` which when written triggers
the writing into the eeprom.

The subset of parameters is currently limited at parameters of devices
`mwc`, `hmc6300` and `hmc6301`. Context attributes and other devices`
parameters will follow.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>