libraries: mqtt: Fix data types
Usually, the payload field of the mqtt_message struct is used in order
to pass "string" data (either a static char* or a buffer formatted
using sprintf). char* and int8_t* may not be equivalent, since on
different CPU architectures char might be defined as either signed or
unsigned, thus a compilation warning is generated. Avoid the need to
cast by using the expected data type.

Apply similar changes for other fields.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>