spi-engine: de-pointerize the dma_flags field in init params
The 'dma_flags' field in the spi_engine_offload_init_param struct is a
pointer to an uint32_t.
This seems a bit much; if this field is 0 (not initialized), the default
behavior (or CYCLIC DMA) is maintained.
It mostly looks like un-necessary reference/dereference.

Also, there are only 2 projects that use this field.
Changing them to the new behavior looks simple.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>