drivers: ad7606: split pre-enable + post-disable functions
For IIO, enabling and disabling the PWM causes interruptions in the sample
captures.
In the basic-example, since we do one capture (and then stop), it's not so
problematic.
But in iio-oscilloscope it's clearly vision that the captures are
interrupted.

So, we just need to implement for hooks to enable PWM before a buffer
capture and disable it after that.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>