projects: eval-ade7753: example project
Example project for ADE7753. Tested using the EVAL-ADE7753.

This is a basic example on how to use the ADE7753 driver.
The platform used is Maxim and the MCU board is the AD-APARD32690-SL.
The example uses SPI communication and prints on the serial port
the rms values for the 2 channels, the temperature and period.
The other measurements are saved in strucutures.

Signed-off-by: Radu Etz <radu.etz@analog.com>